refactor(platform): extract checkout and auto-recharge buttons to own files
Move SaveAutoRechargeButton and CheckoutButton into dedicated files to
reduce billing-dialog.tsx size and narrow the oxlint-disable scope.
